Tweak graphql support
