+++ test/github/graphql.test.ts
@@ -69,7 +69,7 @@ describe('github/graphql', () => {
     })
 
     test('raises errors', async () => {
-      const response = { 'data': null, 'errors': [{ 'message': 'Unexpected end of document' }] }
+      const response = { 'data': 'some data', 'errors': [{ 'message': 'Unexpected end of document' }] }
 
       nock('https://api.github.com').post('/graphql', { query })
         .reply(200, response)
@@ -82,10 +82,11 @@ describe('github/graphql', () => {
       }
 
       expect(thrownError).not.toBeUndefined()
+      expect(thrownError.name).toEqual('GraphQLQueryError')
       expect(thrownError.toString()).toContain('Unexpected end of document')
       expect(thrownError.query).toEqual(query)
       expect(thrownError.errors).toEqual(response.errors)
-      expect(thrownError.data).toEqual(response.data)
+      expect(thrownError.data).toEqual('some data')
     })
   })
